blockchain-analyzer/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
networkx==3.1
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
web3==6.11.1
plotly==5.18.0
dash==2.14.2
dash-cytoscape==1.0.0
